seperated files and update readme
