+++ src/index.js
@@ -1,41 +1,10 @@
-//test code
-const printCorrectAnswerForTest = computerInputNumbers => console.log('hello world! correctAnswer is ', computerInputNumbers);
-
-//variables
-const correctAnswerMessage = "🎉정답을 맞추셨습니다!🎉";
-const getWrongAnswerMessage = ({strike, ball}) => (!strike && !ball) ? '낫싱' : ((ball ? `${ball}볼 ` : '') + (strike ? `${strike}스트라이크` : ''));
-
-//functions to execute baseball game
-const checkValidNumber = inputs => {
-  if(/[^1-9]+/g.test(inputs)) return {ok: false, msg: '1-9의 각 다른 숫자 3개를 공백 없이 입력하세요'};
-  if(inputs.length !== 3) return {ok: false, msg: '숫자는 3개만 입력가능합니다.'};
-  if(inputs.charAt(0) === inputs.charAt(1) || inputs.charAt(0) === inputs.charAt(2) || inputs.charAt(1) === inputs.charAt(2)) return {ok: false, msg: '같은 숫자는 여러번 올 수 없습니다.'};
-  return {ok: true, msg: '올바른 입력입니다!'};
-}
-
-const createRandomNumber = () => {
-  let randomNum = String(Math.floor(Math.random()*999 + 1));
-  return checkValidNumber(randomNum).ok ? randomNum : createRandomNumber();
-}
-
-const compareAnswersAndGetResult = (computerInputNumbers, userInputNumbers) => {
-  if(computerInputNumbers === userInputNumbers) return {ok: true};
-  const computerInputNumbersArr = computerInputNumbers.split('');
-  const userInputNumbersArr = userInputNumbers.split('');
-  let strike = 0, ball = 0;
-
-  //check answer
-  for(let i=0; i<3; i++) {
-    const index = computerInputNumbersArr.indexOf(userInputNumbersArr[i]);
-    strike = index === i ? strike + 1 : strike;
-    ball = index !== i && index !== -1 ? ball + 1 : ball;
-  }
-
-  return {
-    ball,
-    strike
-  }
-}
+import {
+  createRandomNumber,
+  checkValidNumber,
+  compareAnswersAndGetResult 
+} from '/src/lib/gameFunctions.js';
+import { correctAnswerMessage, getWrongAnswerMessage } from '/src/lib/messages.js';
+import { printCorrectAnswerForTest } from '/src/lib/test.js';
 
 const gameTurn = document.getElementById('app');
 const userInput = gameTurn.querySelector('#user-input');
@@ -107,7 +76,6 @@ const playGame = e => {
   if(baseballGame.isFinished) return alert('이미 정답을 맞히셨습니다!');
 
   checkUserInput.ok ? baseballGame.play(baseballGame.answer, userInput.value) : alert(checkUserInput.msg);
-  //console.log(baseballGame.result);
 }
 
 userInputButton.addEventListener('click', playGame);
