+++ src/lib/test.js
@@ -1,0 +1,1 @@
+export const printCorrectAnswerForTest = computerInputNumbers => console.log('hello world! correctAnswer is ', computerInputNumbers);
